+++ NVActivityIndicatorView/NVActivityIndicatorPresenter.swift
@@ -53,6 +53,9 @@ public final class ActivityData {
     /// Minimum display time of UI blocker.
     let minimumDisplayTime: Int
     
+    /// Background color of the UI blocker
+    let backgroundColor: UIColor
+    
     /**
      Create information package used to display UI blocker.
      
@@ -76,7 +79,8 @@ public final class ActivityData {
                 color: UIColor? = nil,
                 padding: CGFloat? = nil,
                 displayTimeThreshold: Int? = nil,
-                minimumDisplayTime: Int? = nil) {
+                minimumDisplayTime: Int? = nil,
+                backgroundColor: UIColor? = nil) {
         self.size = size ?? NVActivityIndicatorView.DEFAULT_BLOCKER_SIZE
         self.message = message ?? NVActivityIndicatorView.DEFAULT_BLOCKER_MESSAGE
         self.messageFont = messageFont ?? NVActivityIndicatorView.DEFAULT_BLOCKER_MESSAGE_FONT
@@ -85,6 +89,7 @@ public final class ActivityData {
         self.padding = padding ?? NVActivityIndicatorView.DEFAULT_PADDING
         self.displayTimeThreshold = displayTimeThreshold ?? NVActivityIndicatorView.DEFAULT_BLOCKER_DISPLAY_TIME_THRESHOLD
         self.minimumDisplayTime = minimumDisplayTime ?? NVActivityIndicatorView.DEFAULT_BLOCKER_MINIMUM_DISPLAY_TIME
+        self.backgroundColor = backgroundColor ?? NVActivityIndicatorView.DEFAULT_BLOCKER_BACKGROUND_COLOR
     }
 }
 
@@ -94,8 +99,16 @@ public final class NVActivityIndicatorPresenter {
     private var hideTimer: Timer?
     private var isStopAnimatingCalled = false
     private let restorationIdentifier = "NVActivityIndicatorViewContainer"
-    
-    
+  
+    private var activitySize = NVActivityIndicatorView.DEFAULT_BLOCKER_SIZE
+
+    private let activityLabel: UILabel = {
+        let activityLabel = UILabel()
+        activityLabel.textAlignment = .center
+        activityLabel.numberOfLines = 0
+        return activityLabel
+    }()
+
     /// Shared instance of `NVActivityIndicatorPresenter`.
     public static let sharedInstance = NVActivityIndicatorPresenter()
     
@@ -122,9 +135,32 @@ public final class NVActivityIndicatorPresenter {
         guard hideTimer == nil else { return }
         hide()
     }
-    
+
+    /// Set message displayed under activity indicator view.
+    ///
+    /// - Parameter message: message displayed under activity indicator view.
+    public final func setMessage(_ message: String?) {
+        activityLabel.text = message
+        
+        guard let message = message, !message.isEmpty else {
+            activityLabel.frame.size = CGSize.zero;
+            return
+        }
+        let screenSize = UIScreen.main.bounds
+        
+        activityLabel.frame.size = NSString(string: message).boundingRect(
+            with: CGSize(width: screenSize.width - 16.0, height: CGFloat.greatestFiniteMagnitude),
+            options: .usesLineFragmentOrigin,
+            attributes: [NSFontAttributeName: activityLabel.font],
+            context: nil).size
+        
+        activityLabel.center = CGPoint(
+            x: screenSize.width / 2.0,
+            y: (screenSize.height / 2.0) + activitySize.height + (activityLabel.frame.height / 2) + 8.0)
+    }
+  
     // MARK: - Timer events
-    
+  
     @objc private func showTimerFired(_ timer: Timer) {
         guard let activityData = timer.userInfo as? ActivityData else { return }
         show(with: activityData)
@@ -137,18 +173,19 @@ public final class NVActivityIndicatorPresenter {
             hide()
         }
     }
-    
+
     // MARK: - Helpers
     
     private func show(with activityData: ActivityData) {
         let activityContainer: UIView = UIView(frame: UIScreen.main.bounds)
         
-        activityContainer.backgroundColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.5)
+        activityContainer.backgroundColor = activityData.backgroundColor
         activityContainer.restorationIdentifier = restorationIdentifier
         
-        let actualSize = activityData.size
+        activitySize = activityData.size
+
         let activityIndicatorView = NVActivityIndicatorView(
-            frame: CGRect(x: 0, y: 0, width: actualSize.width, height: actualSize.height),
+            frame: CGRect(x: 0, y: 0, width: activitySize.width, height: activitySize.height),
             type: activityData.type,
             color: activityData.color,
             padding: activityData.padding)
@@ -156,27 +193,12 @@ public final class NVActivityIndicatorPresenter {
         activityIndicatorView.center = activityContainer.center
         activityIndicatorView.startAnimating()
         activityContainer.addSubview(activityIndicatorView)
-        
-        if let message = activityData.message , !message.isEmpty {
-            let label = UILabel()
-            
-            label.textAlignment = .center
-            label.text = message
-            label.font = activityData.messageFont
-            label.textColor = activityIndicatorView.color
-            label.numberOfLines = 0
-            label.sizeToFit()
-            if label.bounds.size.width > activityContainer.bounds.size.width {
-                let maxWidth = activityContainer.bounds.size.width - 16
-                
-                label.bounds.size = NSString(string: message).boundingRect(with: CGSize(width: maxWidth, height: CGFloat.greatestFiniteMagnitude), options: .usesLineFragmentOrigin, attributes: [NSFontAttributeName: label.font], context: nil).size
-            }
-            label.center = CGPoint(
-                x: activityIndicatorView.center.x,
-                y: activityIndicatorView.center.y + actualSize.height + label.bounds.size.height / 2 + 8)
-            activityContainer.addSubview(label)
-        }
-        
+
+        activityLabel.font = activityData.messageFont
+        activityLabel.textColor = activityIndicatorView.color
+        setMessage(activityData.message)
+        activityContainer.addSubview(activityLabel)
+      
         hideTimer = scheduledTimer(activityData.minimumDisplayTime, selector: #selector(hideTimerFired(_:)), data: nil)
         guard let keyWindow = UIApplication.shared.keyWindow else { return }
         keyWindow.addSubview(activityContainer)